added cli structure
