+++ cli/group.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"fmt"
@@ -8,33 +8,6 @@ import (
 
 	"github.com/spf13/cobra"
 )
-
-var deviceCmd = &cobra.Command{
-	Use:   "device",
-	Short: "Wake a device by name",
-	RunE: func(cmd *cobra.Command, args []string) error {
-		name, _ := cmd.Flags().GetString("name")
-		if name == "" {
-			return cmd.Help()
-		}
-
-		cfg := config.ReadConfig()
-		for _, d := range cfg.Devices {
-			if d.DeviceName == name {
-				if err := wol.WakeDevice(d.MacAddress); err != nil {
-					fmt.Fprintln(os.Stderr, err)
-					os.Exit(1)
-				}
-				fmt.Printf("Waking device %q (%s)\n", d.DeviceName, d.MacAddress)
-				return nil
-			}
-		}
-
-		fmt.Fprintf(os.Stderr, "Device not found: %q\n", name)
-		os.Exit(1)
-		return nil
-	},
-}
 
 var groupCmd = &cobra.Command{
 	Use:   "group",
@@ -76,6 +49,5 @@ var groupCmd = &cobra.Command{
 }
 
 func init() {
-	deviceCmd.Flags().StringP("name", "n", "", "name of the device to wake")
 	groupCmd.Flags().StringP("name", "n", "", "name of the group to wake")
 }
